+++ threadmill-store-redis/src/test/java/com/hemju/threadmill/store/redis/RedisKeysTest.java
@@ -41,7 +41,10 @@ class RedisKeysTest {
         RedisKeys.concurrencyWorkflows("project:42"),
         RedisKeys.concurrencyWorkflowCounts("project:42"),
         RedisKeys.queueKeys("default"),
-        RedisKeys.queueUnkeyed("default"));
+        RedisKeys.queueUnkeyed("default"),
+        RedisKeys.queueEnqueuedAt("default"),
+        RedisKeys.QUEUE_ENQUEUED_AT_LAYOUT,
+        RedisKeys.nodeLayout(NodeId.newId()));
 
     assertThat(keys).allSatisfy(key -> assertThat(SlotHash.getSlot(key)).as(key).isEqualTo(slot));
   }